moved load_source into multiphonon.sqe.__init__.py
