+++ src/multiphonon/sqe/__init__.py
@@ -3,8 +3,22 @@
 # Jiao Lin
 
 
+import importlib.machinery
+import importlib.util
+
 import histogram as H
 import numpy as np
+
+
+def load_source(modname, filename):
+    loader = importlib.machinery.SourceFileLoader(modname, filename)
+    spec = importlib.util.spec_from_file_location(modname, filename, loader=loader)
+    module = importlib.util.module_from_spec(spec)
+    # The module is always executed and not cached in sys.modules.
+    # Uncomment the following line to cache the module.
+    # sys.modules[module.__name__] = module
+    loader.exec_module(module)
+    return module
 
 
 def _conv_unit_label(un):
