+++ tests/redutils_TestCase.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python
 #
 
-import imp
 import os
 import tempfile
 import unittest
 
 import pytest
+from multiphonon.sqe import load_source
 
 # pytestmark = pytest.mark.skipif(False, reason="only run mannually")
 pytestmark = pytest.mark.needs_mantid
@@ -16,7 +16,7 @@ interactive = False
 here = os.path.dirname(__file__)
 datadir = os.path.join(here, "data")
 
-dataurls = imp.load_source("dataurls", os.path.join(datadir, "dataurls.py"))
+dataurls = load_source("dataurls", os.path.join(datadir, "dataurls.py"))
 
 
 class TestCase(unittest.TestCase):
